wip: refactor(agent):
flatten connection and deployment namespaces
