+++ agent/agent.test.ts
@@ -3,26 +3,8 @@
     along with this program.  If not, see <http://www.gnu.org/licenses/>. **/
 import { Suite } from '@hackbg/ensuite'
 export default new Suite([
-  ['base',   ()=>import('./base.test')],
-  ['connection',  ()=>import('./connection.test')],
-  //['client', ()=>import('./client.test')],
-  ['code',   ()=>import('./code.test')],
-  ['deploy', ()=>import('./deploy.test')],
-  ['devnet', ()=>import('./devnet.test')],
-  ['store',  ()=>import('./store.test')],
-  ['token',  ()=>import('./token.test')]
+  ['base',       ()=>import('./base.test')],
+  ['connection', ()=>import('./connec.test')],
+  ['deploy',     ()=>import('./deploy.test')],
+  ['token',      ()=>import('./token.test')]
 ])
-
-import type { Connection, Devnet } from '.'
-import { Fee } from './token'
-export function testChainSupport <
-  A extends typeof Connection, D extends Devnet<A>
-> (
-  Connection: A,
-  devnet: D,
-  token:  string
-) {
-  const sendFee   = new Fee( "1000000", "uknow")
-  const uploadFee = new Fee("10000000", "uknow")
-  const initFee   = new Fee("10000000", "uknow")
-}
