+++ agent/deploy.test.ts
@@ -1,23 +1,36 @@
 /** Fadroma. Copyright (C) 2023 Hack.bg. License: GNU AGPLv3 or custom.
     You should have received a copy of the GNU Affero General Public License
     along with this program.  If not, see <http://www.gnu.org/licenses/>. **/
-import assert from 'node:assert'
-import { Console } from './base'
-import { Deployment, ContractInstance } from './deploy'
-import { ContractClient } from './client'
+import assert, { deepEqual, rejects } from 'node:assert'
+import { Contract } from './connec'
 import * as Stub from './stub'
-import { CompiledCode } from './code'
+import {
+  Deployment,
+  ContractCode,
+  ContractInstance,
+  SourceCode,
+  RustSourceCode,
+  CompiledCode,
+  UploadedCode,
+  UploadStore,
+  DeployStore
+} from './deploy'
 
 import { Suite } from '@hackbg/ensuite'
 export default new Suite([
   ['units',      testDeploymentUnits],
   ['deployment', testDeployment],
+  ['upload',     testUploadStore],
+  ['deploy',     testDeployStore],
+  ['compiler',   testCodeCompiler],
+  ['units',      testCodeUnits],
+  ['contract',   testCodeContract]
 ])
 
 export async function testDeploymentUnits () {
   const contract = new ContractInstance({ address: 'present' })
   assert.equal(await contract.deploy(), contract)
-  assert(contract.connect(new Stub.Agent()) instanceof ContractClient)
+  assert(contract.connect(new Stub.Connection()) instanceof Contract)
 
   assert.rejects(()=>new ContractInstance({
     uploaded: { codeId: 123 } as any,
@@ -72,7 +85,7 @@ export async function testDeployment () {
   }
 
   await new MyDeployment().contract1.deploy({
-    deployer: new Stub.Agent()
+    deployer: new Stub.Connection()
   })
 
   new MyDeployment().contract1.serialize()
@@ -85,12 +98,138 @@ export async function testDeployment () {
 
   await new MyDeployment().upload({
     compiler: new Stub.Compiler(),
-    uploader: new Stub.Agent(),
+    uploader: new Stub.Connection(),
   })
 
   await new MyDeployment().deploy({
     compiler: new Stub.Compiler(),
-    uploader: new Stub.Agent(),
-    deployer: new Stub.Agent(),
+    uploader: new Stub.Connection(),
+    deployer: new Stub.Connection(),
+  })
+}
+
+export async function testUploadStore () {
+  const uploadStore = new UploadStore()
+  assert.equal(uploadStore.get('name'), undefined)
+  assert.equal(uploadStore.set('name', {}), uploadStore)
+  assert.throws(()=>uploadStore.set('foo', { codeHash: 'bar' }))
+  assert(uploadStore.get('name') instanceof UploadedCode)
+}
+
+export async function testDeployStore () {
+  const deployStore = new DeployStore()
+  assert.equal(deployStore.get('name'), undefined)
+  const deployment = new Deployment({ name: 'foo' })
+  assert.equal(deployStore.set('name', deployment), deployStore)
+  assert.deepEqual(deployStore.get('name'), deployment.serialize())
+}
+export async function testCodeContract () {
+  //const contract1 = new ContractCode({
+    //source:   new SourceCode(),
+    //compiled: new CompiledCode(),
+    //uploaded: new UploadedCode()
+  //})
+  //assert(contract1.source instanceof SourceCode)
+  //assert(contract1.compiled instanceof CompiledCode)
+  //assert(contract1.uploaded instanceof UploadedCode)
+  //// can't compile missing code
+  //rejects(()=>contract1.compile())
+  //const validSource = new class extends SourceCode { isValid () { return true } }
+  //const invalidSource = new class extends SourceCode { isValid () { return false } }
+  //const brokenCompiler: any = { build: () => Promise.resolve({ isValid: () => false }) }
+  //rejects(()=>new ContractCode({source: validSource}).compile({compiler: brokenCompiler}))
+  //rejects(()=>new ContractCode({source: invalidSource}).compile({compiler: new Stub.Compiler()}))
+  //assert(new ContractCode({ source: validSource }).compile({ compiler: new Stub.Compiler() }))
+  //// can't upload missing code
+  //rejects(()=>contract1.upload())
+  //rejects(()=>contract1.upload({uploader: new Stub.Connection()}))
+  //rejects(()=>contract1.upload({uploader: {upload: () => Promise.resolve({ isValid: () => false })} as any}))
+  //assert(contract1.source[Symbol.toStringTag] || true)
+  //assert(contract1.compiled[Symbol.toStringTag] || true)
+  ////assert(contract1.uploaded[Symbol.toStringTag])
+  ////assert(contract1.instance[Symbol.toStringTag])
+  //rejects(()=>new CompiledCode().fetch())
+  //rejects(()=>new CompiledCode({ codePath: '' }).fetch())
+  //rejects(()=>new CompiledCode({ codePath: new URL('', 'file:') }).fetch())
+  //rejects(()=>new CompiledCode({ codePath: new URL('http://foo.bar') }).fetch())
+  //rejects(()=>new CompiledCode({ codePath: 0 as any }).fetch())
+}
+
+export async function testCodeCompiler () {
+  assert((await new Stub.Compiler().build('')) instanceof CompiledCode)
+  assert((await new Stub.Compiler().buildMany([{}]))[0] instanceof CompiledCode)
+}
+
+export async function testCodeUnits () {
+
+  rejects(()=>new ContractCode({ }).compile())
+
+  rejects(()=>new ContractCode({ compiler: new Stub.Compiler() }).compile())
+
+  const source1 = new SourceCode()
+  assert(source1[Symbol.toStringTag])
+  deepEqual(source1.serialize(), {
+    sourceOrigin: undefined,
+    sourceRef:    undefined,
+    sourcePath:   undefined,
+    sourceDirty:  undefined,
+  })
+  assert(!source1.canFetch)
+  assert(!source1.canCompile)
+  source1.sourceOrigin = 'foo'
+  assert(source1.canFetch)
+  assert(source1.canCompile)
+  source1.sourceOrigin = undefined
+  source1.sourcePath = 'foo'
+  assert(!source1.canFetch)
+  assert(source1.canCompile)
+
+  rejects(()=>new ContractCode({ source: source1 }).compile())
+
+  assert(
+    await new ContractCode({ source: source1, compiler: new Stub.Compiler() }).compile()
+    instanceof CompiledCode
+  )
+
+  const rustSource1 = new RustSourceCode()
+  assert(rustSource1[Symbol.toStringTag])
+  deepEqual(rustSource1.serialize(), {
+    sourceOrigin:   undefined,
+    sourceRef:      undefined,
+    sourcePath:     undefined,
+    sourceDirty:    undefined,
+    cargoToml:      undefined,
+    cargoWorkspace: undefined,
+    cargoCrate:     undefined,
+    cargoFeatures:  undefined,
+  })
+  assert(!rustSource1.canFetch)
+  assert(!rustSource1.canCompile)
+  rustSource1.sourceOrigin = 'foo'
+  assert(rustSource1.canFetch)
+  assert(!rustSource1.canCompile)
+  rustSource1.sourceOrigin = undefined
+  rustSource1.sourcePath = 'foo'
+  assert(!rustSource1.canFetch)
+  assert(!rustSource1.canCompile)
+  rustSource1.cargoToml = 'foo'
+  assert(rustSource1.canCompile)
+  rustSource1.canFetch
+  rustSource1.canCompileInfo
+
+  const compiled1 = new CompiledCode()
+  deepEqual(compiled1.serialize(), {
+    codeHash: undefined,
+    codePath: undefined,
+  })
+
+  const uploaded1 = new UploadedCode()
+  deepEqual(uploaded1.serialize(), {
+    codeHash:  undefined,
+    chainId:   undefined,
+    codeId:    undefined,
+    uploadBy:  undefined,
+    uploadTx:  undefined,
+    uploadGas: undefined
   })
 }
